requirements.txt
sqlalchemy
pymysql
matplotlib
pandas
openpyxl
reportlab